pyproject.toml
[tool.poetry]
name = "eiproject"
version = "0.1.0"
description = ""
authors = ["Christoph Graf <31027604+chrigraf@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "3.10.6"
geopandas = "0.12.2"
shapely = "1.8.5"
matplotlib = "3.6.2"
requests = "2.28.1"
pygeos = "0.13"
openpyxl = "^3.1.2"

[tool.poetry.dev-dependencies]

[tool.poetry.group.dev.dependencies]
black = {version = "^23.1a1", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
